# doc/api.yaml
openapi: 3.0.3
servers:
  - url: http://localhost:3000
    description: Local server
info:
  title: WASAPhoto API
  version: 1.0.0
...
paths:
  /session:
    post:
      tags:
        - login
      operationId: doLogin
      summary: Logs in the user
      description: >-
        If the user does not exist it will be created and an identifier is
        returned. If the user exists the user identifier is returned.
      requestBody:
        description: User details
        content:
          application/json:
            schema: {$ref: "#/components/schemas/user_name"}
        required: true
      responses:
        "201":
          description: User log-in action successfull
          content:
            application/json:
              schema:
                description: userName identifier
                type: object
                properties:
                  identifier:
                    description: User identifier
                    type: integer
        "400": {$ref: '#/components/responses/BadRequest'}
        "401": {$ref: '#/components/responses/Unauthorized'}
        "500": {$ref: '#/components/responses/InternalServerError'}
      security:
        - HTTPBearerAuth: []
...
  /user/{uid}/photos/{photoid}:
    parameters:
    - name: uid
      in: path
      required: true
      description: this is the user id
      schema:
        type: integer
        description: e.g., /user/1234

    - name: photoid
      in: path
      required: true
      description: this is the photo id that the main user wants to delete
      schema:
        type: integer
        description: e.g., /user/1234/photos/1357
    
    post:
      tags:
        - Comment a photo
      operationId: commentPhoto
      summary: Comment a photo
      description: |
        The user {uid} wants to comment the photo {photoid}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              description: The user need to insert the comment that he want to post
              type: object
              properties:
                text:
                  description: Comment text
                  type: string
                  minLength: 3
                  maxLength: 300
                  pattern: ^.*?$
      responses:
        "201":
          description: User {uid} has commented the photo {photoid}
          content: 
            application/json: 
              schema:
                description: object for comment information 
                type: object
                properties:
                  status:
                    description: Status of the opeartion
                    type: string
                    minLength: 3
                    maxLength: 20
                  message:
                    description: Message of the operation
                    type: string
                    minLength: 3
                    maxLength: 20
                  data: {$ref: '#/components/schemas/object_comment'}
        "400": {$ref: '#/components/responses/BadRequest'}
        "500": {$ref: '#/components/responses/InternalServerError'}

      security:
        - HTTPBearerAuth: []

    delete:
      tags: 
        - Delete photo
      operationId: deletePhoto
      summary: Delete a photo
      description: |
        The user {uid} wants to delete a photo
      responses: 
        "204":
          description: Photo deleted successfully
        "400": {$ref: '#/components/responses/BadRequest'}
        "500": {$ref: '#/components/responses/InternalServerError'}

      security:
        - HTTPBearerAuth: []
...
  /user/{uid}/follow/{fluid}:
    parameters:
    - name: uid
      in: path
      required: true
      description: this is the user id
      schema:
        type: integer
        description: e.g., /user/1234

    - name: fluid
      in: path
      required: true
      description: this is the user id that the main user wants to follow or unfollow
      schema:
        type: integer
        description: e.g., /user/1234/follow/1357
    
    put:
      tags: 
        - Follow User
      operationId: followUser
      summary: Follow an user
      description: |
        The user {uid} wants to follow the user {fluid}
      responses:
        "201":
          description: User {uid} is now following user {fluid}
          content: 
            application/json: 
              schema:
                description: object for comment information
                type: object
                properties:
                  status:
                    description: Status of the opeartion
                    type: string
                    minLength: 3
                    maxLength: 20
                  message:
                    description: Message of the operation
                    type: string
                    minLength: 3
                    maxLength: 50
                  data: {$ref: '#/components/schemas/object_followed'}
        "400": {$ref: '#/components/responses/BadRequest'}
        "500": {$ref: '#/components/responses/InternalServerError'}

      security:
        - HTTPBearerAuth: []

    delete:
      tags: 
        - Unfollow an user
      operationId: unfollowUser
      summary: Unfollow an user
      description: |
        The user {uid} wants to unfollow the user {fluid}
      responses:
        "204":
          description: User {uid} is not following anymore user {fluid}

      security:
        - HTTPBearerAuth: []
  /users/{uid}/follows/{fluid}:
    parameters:
    - name: uid
      in: path
      required: true
      description: this is the user id
      schema:
        type: integer
        description: e.g., /user/1234

    - name: fluid
      in: path
      required: true
      description: this is the user id that the main user wants to check if he is following
      schema:
        type: integer
        description: e.g., /user/1234/follow/1357
    
    get:
      tags: 
        - Get Followers
      operationId: checkFollower
      summary: check Follower
      description: |
        Get the object profile of the follower if the user {uid} is following the user {fluid}
      responses: 
        "200":
          description: Get an object profile by the ID
          content:
            application/json: 
              schema:
               description: boolean value that indicates if the user {uid} is following the user {fluid}
               type: boolean
        "400": {$ref: '#/components/responses/BadRequest'}
        "500": {$ref: '#/components/responses/InternalServerError'} 
      security:
        - HTTPBearerAuth: []
  /user/{uid}/ban/{banuid}:
    parameters:
    - name: uid
      in: path
      required: true
      description: this is the user id
      schema:
        type: integer
        description: e.g., /user/1234

    - name: banuid
      in: path
      required: true
      description: this is the user id that the main user wants to ban or unban
      schema:
        type: integer
        description: e.g., /user/1234/ban/1357
    
    post:
      tags: 
        - Ban an user
      operationId: banUser
      summary: Ban an user
      description: |
        The user {uid} wants to ban the user {banuid}
      responses:
        "201":
          description: User {banuid} has been banned from user {uid}
          content: 
            application/json: 
              schema:
                description: object for ban information
                type: object
                properties:
                  status:
                    description: Status of the opeartion
                    type: string
                    minLength: 3
                    maxLength: 20
                  message:
                    description: Message of the operation
                    type: string
                    minLength: 3
                    maxLength: 50
                  data: {$ref: '#/components/schemas/object_banned'}
        "400": {$ref: '#/components/responses/BadRequest'}
        "500": {$ref: '#/components/responses/InternalServerError'}
      
      security:
        - HTTPBearerAuth: []

    delete:
      tags: 
        - Unban an user
      operationId: unbanUser
      summary: Unban an user
      description: |
        The user {uid} wants to unban the user {fluid}
      responses:
        "204":
          description: User {banuid} has been unbanned from user {uid}
        "400": {$ref: '#/components/responses/BadRequest'}
        "500": {$ref: '#/components/responses/InternalServerError'}
      
      security:
        - HTTPBearerAuth: []
  /users/{uid}/bans/{banuid}:
    parameters:
      - name: uid
        in: path
        required: true
        description: this is the user id
        schema:
          type: integer
          description: e.g., /user/1234
      
      - name: banuid
        in: path
        required: true
        description: this is the user id that the main user wants to check if he is banned
        schema:
          type: integer
          description: e.g., /user/1234/ban/1357
      
    get:
      tags: 
        - Get Banned
      operationId: checkBanned
      summary: check Banned
      description: |
        Get the object profile of the banned user if the user {uid} has banned the user {banuid}
      responses: 
        "200":
          description: Get an object profile by the ID
          content:
            application/json: 
              schema:
               description: boolean value that indicates if the user {uid} has banned the user {banuid}
               type: boolean
        "400": {$ref: '#/components/responses/BadRequest'}
        "500": {$ref: '#/components/responses/InternalServerError'} 
      security:
        - HTTPBearerAuth: []
...
components:
  responses:
    Unauthorized:
      description: The client provides no credentials or invalid credentials
    BadRequest:
      description: The request was not compliant with the documentation (eg. missing fields, etc)
    InternalServerError:
      description: The server encountered an internal error. Further info in server logs
  securitySchemes: 
    HTTPBearerAuth:
      type: http
      scheme: bearer
  schemas:
   user_name:
      description: userName object schema
      type: object
      properties:
        name:
          description: User name structure
          type: string
          example: Vincenzo
          pattern: ^.*?$
          minLength: 3
          maxLength: 16
...
   object_comment:
      description: object for comment information
      type: object
      properties:
        commentId:
          description: Comment identifier
          type: integer
        text:
          description: Comment text
          type: string
          minLength: 3
          maxLength: 300
        userId:
          description: User identifier
          type: integer
   object_followed:
      description: object for followed user
      type: object
      properties:
        followerId:
           description: User identifier
           type: integer
        followingId:
           description: Followed user identifier
           type: integer
   object_banned:
        description: object for banned user
        type: object
        properties:
          userId:
            description: User identifier
            type: integer
          bannedId:
            description: Banned user identifier
            type: integer
...